Move meetup event components into events folder
