+++ client/components/events/meetupEventListItem.js
